Factor out reactivation related logic into its own module.
